# PERF: avoid copies if possible in fill_binop