@jbrockmendel 2020-01-25
> Not being super familiar here, is there a way to avoid copy without having to special case for the mask presence?

Not sure I understand the question.  Can you give an example?