@gfyoung 2020-01-25
> Not sure I understand the question.

I believe @WillAyd is trying to determine if there is a faster way to figure out if we need to `copy` or not.  `.any` is still `O(n)`, so the average run-time is theoretically unchanged with your diff.