@jbrockmendel 2020-01-25
> determine if there is a faster way to figure out if we need to .copy or not

Well, we could pre-check for dtypes that can't hold NAs and short-circuit in those cases.

Other than that, we might be able to do 1 O(n) pass instead of 2-3 if we did it all in cython, but i dont think thats worthwhile at this point